chore(team): split conversation manager tests
